[수정] 메인페이지 컴포넌트 분리, 스타일 수정
