--- src/pages/Main/Gym.jsx
@@ -1,3 +1,0 @@
-export default function Gym() {
-  return <h1 className="text-3xl">헬스장 상세 정보 페이지입니다</h1>;
-}
