+++ src/pages/Main/Gym/Gym.jsx
@@ -1,0 +1,10 @@
+import AppMeta from '@/components/AppMeta';
+
+export default function Gym() {
+  return (
+    <>
+      <AppMeta title="헬스장 페이지" description="헬스장 페이지입니다." />
+      <h1 className="text-3xl">헬스장 상세 정보 페이지입니다</h1>;
+    </>
+  );
+}
