--- src/pages/Main/Map.jsx
@@ -1,3 +1,0 @@
-export default function Map() {
-  return <h1 className="text-3xl">지도 페이지입니다</h1>;
-}
