+++ src/pages/Main/Map/Map.jsx
@@ -1,0 +1,10 @@
+import AppMeta from '@/components/AppMeta';
+
+export default function Map() {
+  return (
+    <>
+      <AppMeta title="지도 페이지" description="지도 페이지입니다." />
+      <h1 className="text-3xl">지도 페이지입니다</h1>;
+    </>
+  );
+}
